requirements.txt
torch>=2.4.0
unsloth
transformers
trl
pillow
